feat: Add Czech quotes from Wikiquote
Adds a list of quotes extracted from Czech instance of Wikiquote
by https://github.com/ceskyDJ/international-quotes-api. As there
are many quotes and some of them are of lower quality, this commit
adds only quotes scored as 90 and above by the linked extractor.
See its README and/or code for details of scoring mechanism.